add "modules-late" service
Some modules, when loaded, trigger hotplugging of services that in turn
need some services from the "default" runlevel (like dbus) or which need
local filesystems to be mounted (including the root filesystem (re)mounted
read / write).

This is especially true for devices like NICs or USB modems, but almost
certainly not limited to them.

For example, in current netifrc a wireless NIC handled by wpa_supplicant
module needs dbus service to be started (if wpa_supplicant was built with
dbus support) and these NICs which are handled by dhcpcd need
/var/lib/dhcpcd/ writable (in practice this most often means a need to
have the root filesystem writable).
We order the normal "modules" service before "fsck" service (and so before
remounting the root fs read / write) so filesystems that are on devices
that need some modules loaded can be checked.

There used to be also a problem with usb_modeswitch needing a write access
to the root filesystem when switching mode of an USB modem (see Gentoo bug
565262). It was possible to work it around that time, but since programs
are clearly expecting a root fs being writable it can be reasonably
expected that similar issues will pop up in the future.

To fix this once for all let's introduce a "modules-late" service started
early in the "default" runlevel which can be used to load modules that have
the requirements described above.
This service is implemented via a slight modification of the existing
"modules" script.