cgroups: enhance support
Add a possibility to:
* Set a cgroup name for a service,

* Set a cgroup name prefix for services that don't have an explicit cgroup
name set by the previous feature.
This allows defining, for example, a root "services" cgroup, under which
children cgroups for particular services are kept,

* Add a service to a controller without needing to set any of its settings,

* Keep a service cgroup when it becomes empty (for example to keep between
service restarts some cgroup settings that were set outside OpenRC).

These functionalities allow configuration of a service cgroup name for v1
controllers and v2 (unified) hierarchy; the service name in the "openrc"
hierarchy is not affected by these settings.

The default values are set in a way so not to result in any change of
services cgroup behavior unless specifically configured in a different way
by an user.